splitLastWord > words | helper file rename
